requirements.txt
flask
requests
python-dotenv==1.0.1